Final touches

+++ sales/api/sales_rest/models.py
@@ -45,4 +45,4 @@ class Sale(models.Model):
         related_name="sales",
         on_delete=models.CASCADE
     )
-    price = models.PositiveSmallIntegerField()
+    price = models.PositiveIntegerField()
